# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  aclmodels.Acl2Scope:
    enum:
    - ""
    - ror
    - cluster
    - project
    - datacenter
    type: string
    x-enum-comments:
      Acl2ScopeRor: ROR
      Acl2ScopeUnknown: unknown
    x-enum-varnames:
    - Acl2ScopeUnknown
    - Acl2ScopeRor
    - Acl2ScopeCluster
    - Acl2ScopeProject
    - Acl2ScopeDatacenter
...
  apicontracts.Datacenter:
    properties:
      apiEndpoint:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/apicontracts.DatacenterLocation'
      name:
        type: string
      provider:
        $ref: '#/definitions/providers.ProviderType'
    type: object
  apicontracts.DatacenterLocation:
    properties:
      country:
        type: string
      id:
        type: string
      region:
        type: string
    type: object
...
  apiresourcecontracts.ResourceAction:
    enum:
    - Add
    - Delete
    - Update
    type: string
    x-enum-varnames:
    - K8sActionAdd
    - K8sActionDelete
    - K8sActionUpdate
...
  apiresourcecontracts.ResourceMetadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      creationTimestamp:
        type: string
      generation:
        type: integer
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      ownerReferences:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceMetadataOwnerReference'
        type: array
      resourceVersion:
        type: string
      uid:
        type: string
    type: object
  apiresourcecontracts.ResourceMetadataOwnerReference:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      name:
        type: string
      uid:
        type: string
    type: object
  apiresourcecontracts.ResourceNode:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/apiresourcecontracts.ResourceMetadata'
      spec:
        $ref: '#/definitions/apiresourcecontracts.ResourceNodeSpec'
      status:
        $ref: '#/definitions/apiresourcecontracts.ResourceNodeStatus'
    type: object
  apiresourcecontracts.ResourceNodeSpec:
    properties:
      podCIDR:
        type: string
      podCIDRs:
        items:
          type: string
        type: array
      providerID:
        type: string
      taints:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceNodeSpecTaints'
        type: array
    type: object
  apiresourcecontracts.ResourceNodeSpecTaints:
    properties:
      effect:
        type: string
      key:
        type: string
    type: object
  apiresourcecontracts.ResourceNodeStatus:
    properties:
      addresses:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceNodeStatusAddresses'
        type: array
      capacity:
        $ref: '#/definitions/apiresourcecontracts.ResourceNodeStatusCapacity'
      conditions:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceNodeStatusConditions'
        type: array
      nodeInfo:
        $ref: '#/definitions/apiresourcecontracts.ResourceNodeStatusNodeinfo'
    type: object
  apiresourcecontracts.ResourceNodeStatusAddresses:
    properties:
      address:
        type: string
      type:
        type: string
    type: object
  apiresourcecontracts.ResourceNodeStatusCapacity:
    properties:
      cpu:
        type: string
      ephemeral-storage:
        type: string
      memory:
        type: string
      pods:
        type: string
    type: object
  apiresourcecontracts.ResourceNodeStatusConditions:
    properties:
      lastHeartbeatTime:
        type: string
      lastTransitionTime:
        type: string
      message:
        type: string
      reason:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  apiresourcecontracts.ResourceNodeStatusNodeinfo:
    properties:
      architecture:
        type: string
      bootID:
        type: string
      containerRuntimeVersion:
        type: string
      kernelVersion:
        type: string
      kubeProxyVersion:
        type: string
      kubeletVersion:
        type: string
      machineID:
        type: string
      operatingSystem:
        type: string
      osImage:
        type: string
      systemUUID:
        type: string
    type: object
  apiresourcecontracts.ResourceOwnerReference:
    properties:
      scope:
        allOf:
        - $ref: '#/definitions/aclmodels.Acl2Scope'
        description: cluster, workspace,...
      subject:
        description: ror id eg clusterId or workspaceName
        type: string
    type: object
  apiresourcecontracts.ResourceUpdateModel:
    properties:
      action:
        $ref: '#/definitions/apiresourcecontracts.ResourceAction'
      apiVersion:
        type: string
      hash:
        type: string
      kind:
        type: string
      owner:
        $ref: '#/definitions/apiresourcecontracts.ResourceOwnerReference'
      resource: {}
      uid:
        type: string
    type: object
...
  providers.ProviderType:
    enum:
    - unknown
    - tanzu
    - aks
    - k3d
    - kind
    type: string
    x-enum-varnames:
    - ProviderTypeUnknown
    - ProviderTypeTanzu
    - ProviderTypeAks
    - ProviderTypeK3d
    - ProviderTypeKind
...
  ror_internal_rorresources.Resource:
    properties:
      api_version:
        type: string
      application:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceApplication'
      appproject:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceAppProject'
      certificate:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceCertificate'
      clustercompliancereport:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceClusterComplianceReport'
      clusterorder:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceClusterOrder'
      configauditreport:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceConfigAuditReport'
      configuration:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceConfiguration'
      daemonset:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceDaemonSet'
      deployment:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceDeployment'
      exposedsecretreport:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceExposedSecretReport'
      ingress:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceIngress'
      ingressclass:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceIngressClass'
      kind:
        type: string
      kubernetescluster:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceKubernetesCluster'
      namespace:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceNamespace'
      node:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceNode'
      persistentvolumeclaim:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourcePersistentVolumeClaim'
      pod:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourcePod'
      policyreport:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourcePolicyReport'
      project:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceProject'
      rbacassessmentreport:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceRbacAssessmentReport'
      replicaset:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceReplicaSet'
      service:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceService'
      statefulset:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceStatefulSet'
      storageclass:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceStorageClass'
      tanzukubernetescluster:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceTanzuKubernetesCluster'
      tanzukubernetesrelease:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceTanzuKubernetesRelease'
      virtualmachineclass:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceVirtualMachineClass'
      virtualmachineclassbinding:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceVirtualMachineClassBinding'
      vulnerabilityreport:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceVulnerabilityReport'
    type: object
  ror_internal_rorresources.ResourceSet:
    properties:
      resources:
        items:
          $ref: '#/definitions/ror_internal_rorresources.Resource'
        type: array
    type: object
...
  ror_internal_rorresources_rortypes.ResourceAppProject:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      rormeta:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceRorMeta'
      spec:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceAppProjectSpec'
    type: object
...
  ror_internal_rorresources_rortypes.ResourceApplication:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      rormeta:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceRorMeta'
      spec:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceApplicationSpec'
      status:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceApplicationStatus'
    type: object
...
  ror_internal_rorresources_rortypes.ResourceCertificate:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      rormeta:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceRorMeta'
      spec:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceCertificateSpec'
      status:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceCertificateStatus'
    type: object
...
  ror_internal_rorresources_rortypes.ResourceClusterComplianceReport:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      rormeta:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceRorMeta'
    type: object
...
  ror_internal_rorresources_rortypes.ResourceRorMeta:
    properties:
      action:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.ResourceAction'
      hash:
        type: string
      internal:
        type: boolean
      lastReported:
        type: string
      ownerref:
        $ref: '#/definitions/ror_internal_rorresources_rortypes.RorResourceOwnerReference'
      version:
        type: string
    type: object
...
info:
  contact:
    name: Privat Sky
    url: http://ror.sky.nhn.no
  description: 'ROR-API, need any help? Go to channel #drift-sdi-devops in norskhelsenett.slack.com
    slack workspace'
  title: Swagger ROR-API
  version: "0.1"
paths:
...
  /v1/datacenters:
    get:
      consumes:
      - application/json
      description: Get datacenters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apicontracts.Datacenter'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get datacenters
      tags:
      - datacenters
    post:
      consumes:
      - application/json
      description: Create datacenter
      parameters:
      - description: Datacenter
        in: body
        name: datacenter
        required: true
        schema:
          $ref: '#/definitions/apicontracts.Datacenter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apicontracts.Datacenter'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Create datacenter
      tags:
      - datacenters
  /v1/datacenters/{datacenterId}:
    put:
      consumes:
      - application/json
      description: Update a datacenter by id
      parameters:
      - description: datacenterId
        in: path
        name: datacenterId
        required: true
        type: string
      - description: Datacenter
        in: body
        name: datacenter
        required: true
        schema:
          $ref: '#/definitions/apicontracts.Datacenter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apicontracts.Datacenter'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Update a datacenter
      tags:
      - datacenters
  /v1/datacenters/{datacenterName}:
    get:
      consumes:
      - application/json
      description: Get datacenter by name
      parameters:
      - description: datacenterName
        in: path
        name: datacenterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.Datacenter'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get datacenter by name
      tags:
      - datacenters
  /v1/datacenters/id/{id}:
    get:
      consumes:
      - application/json
      description: Get datacenter by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.Datacenter'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get datacenter by id
      tags:
      - datacenters
...
  /v1/resources:
    get:
      consumes:
      - application/json
      description: Get a list of resources
      parameters:
      - description: The kind of the owner, currently only support 'Cluster'
        enum:
        - ""
        - ror
        - cluster
        - project
        - datacenter
        in: query
        name: ownerScope
        required: true
        type: string
      - description: The name og the owner
        in: query
        name: ownerSubject
        required: true
        type: string
      - description: ApiVersion
        in: query
        name: apiversion
        required: true
        type: string
      - description: Kind
        in: query
        name: kind
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apiresourcecontracts.ResourceNode'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Registers a  resource
      parameters:
      - description: ResourceUpdate
        in: body
        name: resourcereport
        required: true
        schema:
          $ref: '#/definitions/apiresourcecontracts.ResourceUpdateModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Register  resource
      tags:
      - resources
...
  /v2/resources:
    get:
      consumes:
      - application/json
      description: Get a list of resources
      parameters:
      - description: The kind of the owner, currently only support 'Cluster'
        enum:
        - ""
        - ror
        - cluster
        - project
        - datacenter
        in: query
        name: ownerScope
        required: true
        type: string
      - description: The name og the owner
        in: query
        name: ownerSubject
        required: true
        type: string
      - description: ApiVersion
        in: query
        name: apiversion
        required: true
        type: string
      - description: Kind
        in: query
        name: kind
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apiresourcecontracts.ResourceNode'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Registers a  resource
      parameters:
      - description: ResourceUpdate
        in: body
        name: rorresource
        required: true
        schema:
          $ref: '#/definitions/ror_internal_rorresources.ResourceSet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Register  resource
      tags:
      - resources
...
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
  ApiKey:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"